Changed home-assistant to run in a VM.
This is desired because NIX doesn't have a way, that I'm aware of,
to have a fully working home-assistant. The creators have their own
distro so that was a natual choice.

Not all the files and processes nessiary to have this be declairitive.
Steps done manually or via a bash file.
 - add br0 network into libvirtd
 - download home-assistant-distro-iso
 - bash script to create it. -- Maybe this can be a configuration file?